+++ qiskit_aqua_interfaces/aqua/command_line/command_line.py
@@ -20,7 +20,10 @@ import json
 from collections import OrderedDict
 import textwrap
 import logging
+from qiskit_aqua_interfaces import APP_DEPRECATION_MSG
 from qiskit_aqua_interfaces._extras_require import _check_extra_requires
+
+# pylint: disable=import-outside-toplevel
 
 
 def main():
@@ -103,6 +106,7 @@ def _run():
     with open(args.input) as json_file:
         params = json.load(json_file)
 
+    print(APP_DEPRECATION_MSG)
     ret = run_algorithm(params, None, True)
 
     if args.jo is not None: