+++ qiskit_aqua_interfaces/aqua/user_interface/_aquathread.py
@@ -22,15 +22,23 @@ import io
 import platform
 import os
 import subprocess
+import traceback
 import psutil
 from qiskit_aqua_interfaces.user_interface import GUIProvider
 
 logger = logging.getLogger(__name__)
 
 
+def exception_to_string(excp):
+    """ exception string formatter """
+    stack = traceback.extract_stack()[:-3] + traceback.extract_tb(excp.__traceback__)
+    pretty = traceback.format_list(stack)
+    return ''.join(pretty) + '\n  {} {}'.format(excp.__class__, excp)
+
+
 class AquaThread(threading.Thread):
     """ Aqua Thread """
-    def __init__(self, model, output, queue):
+    def __init__(self, model, output, queue) -> None:
         super(AquaThread, self).__init__(name='Aqua run thread')
         self.model = model
         self._output = output
@@ -61,9 +69,6 @@ class AquaThread(threading.Thread):
         input_file = None
         temp_input = False
         try:
-            algorithms_directory = os.path.dirname(os.path.realpath(__file__))
-            algorithms_directory = os.path.abspath(
-                os.path.join(algorithms_directory, '../command_line'))
             input_file = self.model.get_filename()
             if input_file is None or self.model.is_modified():
                 f_d, input_file = tempfile.mkstemp(suffix='.in')
@@ -72,42 +77,12 @@ class AquaThread(threading.Thread):
                 self.model.save_to_file(input_file)
 
             startupinfo = None
-            process_name = psutil.Process().exe()
-            if not process_name:
-                process_name = 'python'
-            else:
-                if sys.platform == 'win32' and process_name.endswith('pythonw.exe'):
-                    path = os.path.dirname(process_name)
-                    files = [f for f in os.listdir(path) if f != 'pythonw.exe' and f.startswith(
-                        'python') and f.endswith('.exe')]
-                    # sort reverse to have hihre python
-                    # versions first: python3.exe before python2.exe
-                    files = sorted(files, key=str.lower, reverse=True)
-                    new_process = None
-                    for file in files:
-                        proc = os.path.join(path, file)
-                        if os.path.isfile(proc):
-                            # python.exe takes precedence
-                            if file.lower() == 'python.exe':
-                                new_process = proc
-                                break
+            if sys.platform == 'win32':
+                startupinfo = subprocess.STARTUPINFO()
+                startupinfo.dwFlags = subprocess.STARTF_USESHOWWINDOW
+                startupinfo.wShowWindow = subprocess.SW_HIDE
 
-                            # use first found
-                            if new_process is None:
-                                new_process = proc
-
-                    if new_process is not None:
-                        startupinfo = subprocess.STARTUPINFO()
-                        startupinfo.dwFlags = subprocess.STARTF_USESHOWWINDOW
-                        startupinfo.wShowWindow = subprocess.SW_HIDE
-                        process_name = new_process
-
-            if self._output is not None and logger.getEffectiveLevel() == logging.DEBUG:
-                self._output.write('Process: {}\n'.format(process_name))
-
-            self._popen = subprocess.Popen([process_name,
-                                            algorithms_directory,
-                                            input_file],
+            self._popen = subprocess.Popen(['qiskit_aqua_cmd', input_file],
                                            stdin=subprocess.DEVNULL,
                                            stdout=subprocess.PIPE,
                                            stderr=subprocess.STDOUT,
@@ -126,7 +101,7 @@ class AquaThread(threading.Thread):
             self._popen.wait()
         except Exception as ex:  # pylint: disable=broad-except
             if self._output is not None:
-                self._output.write('Process has failed: {}'.format(str(ex)))
+                self._output.write('Process has failed: {}'.format(exception_to_string(ex)))
         finally:
             self._popen = None
             if self._thread_queue is not None: