+++ qiskit_aqua_interfaces/aqua/user_interface/command_line.py
@@ -21,6 +21,8 @@ from qiskit_aqua_interfaces._extras_require import _check_extra_requires
 from qiskit_aqua_interfaces.user_interface._mainview import MainView
 from ._aquaguiprovider import AquaGUIProvider
 
+# pylint: disable=import-outside-toplevel
+
 
 def set_preferences_logging():
     """Update logging setting with latest external packages"""