+++ qiskit_aqua_interfaces/chemistry/user_interface/_model.py
@@ -22,6 +22,8 @@ from ._uipreferences import UIPreferences
 
 logger = logging.getLogger(__name__)
 
+# pylint: disable=import-outside-toplevel
+
 
 class Model(BaseModel):
     """ Chemistry Model """
@@ -100,7 +102,7 @@ class Model(BaseModel):
         return properties_with_substitution
 
     def get_operator_section_names(self):
-        """ get operator sectio  names """
+        """ get operator section names """
         from qiskit.aqua.parser import JSONSchema
         from qiskit.chemistry.parser import InputParser
         from qiskit.chemistry.core import local_chemistry_operators