+++ qiskit_aqua_interfaces/user_interface/_mainview.py
@@ -22,17 +22,19 @@ import tkinter.filedialog as tkfd
 from tkinter import font
 import webbrowser
 import os
-from qiskit_aqua_interfaces import __version__
+from qiskit_aqua_interfaces import __version__, APP_DEPRECATION_MSG
 from ._sectionsview import SectionsView
 from ._sectiontextview import SectionTextView
 from ._threadsafeoutputview import ThreadSafeOutputView
 from ._emptyview import EmptyView
 from ._preferencesdialog import PreferencesDialog
 
+# pylint: disable=import-outside-toplevel
+
 
 class MainView(ttk.Frame):
     """ Main View """
-    def __init__(self, parent, guiprovider):
+    def __init__(self, parent, guiprovider) -> None:
         """Create MainView object."""
         super(MainView, self).__init__(parent)
         self._guiprovider = guiprovider
@@ -235,7 +237,7 @@ class MainView(ttk.Frame):
         # redirect output
         sys.stdout = self._guiprovider.controller.outputview
         sys.stderr = self._guiprovider.controller.outputview
-        # reupdate logging after redirect
+        # update logging after redirect
         self.after(0, self._set_preferences_logging)
 
         self.update_idletasks()
@@ -244,6 +246,8 @@ class MainView(ttk.Frame):
         self._guiprovider.controller._sections_view.show_defaults_button(False)
         self._guiprovider.controller._empty_view.set_toolbar_size(
             self._guiprovider.controller._sections_view.get_toolbar_size())
+
+        self._guiprovider.controller.outputview.write_line(APP_DEPRECATION_MSG)
 
     def _set_preferences_logging(self):
         preferences = self._guiprovider.create_uipreferences()