+++ qiskit_aqua_interfaces/user_interface/_preferencesdialog.py
@@ -21,6 +21,8 @@ import logging
 from ._dialog import Dialog
 from ._credentialsview import CredentialsView
 
+# pylint: disable=import-outside-toplevel
+
 
 class PreferencesDialog(Dialog):
     """ Preferences Dialog """
@@ -33,7 +35,7 @@ class PreferencesDialog(Dialog):
          (logging.NOTSET, logging.getLevelName(logging.NOTSET))]
     )
 
-    def __init__(self, parent, guiprovider):
+    def __init__(self, parent, guiprovider) -> None:
         super(PreferencesDialog, self).__init__(guiprovider.controller, parent, 'Preferences')
         self._guiprovider = guiprovider
         self._credentialsview = None