+++ qiskit_aqua_interfaces/user_interface/base_controller.py
@@ -29,12 +29,14 @@ from ._customwidgets import (EntryPopup, ComboboxPopup, TextPopup)
 
 logger = logging.getLogger(__name__)
 
+# pylint: disable=import-outside-toplevel
+
 
 class BaseController(ABC):
     """Base GUI Controller."""
 
     @abstractmethod
-    def __init__(self, guiprovider, model):
+    def __init__(self, guiprovider, model) -> None:
         self._view = None
         self._guiprovider = guiprovider
         self._model = model
@@ -183,7 +185,9 @@ class BaseController(ABC):
                 self.model.load_file(filename)
             except Exception as ex:  # pylint: disable=broad-except
                 messagebox.showerror("Error", str(ex))
-                ret = False
+                # Only return false if no file or if file not found
+                if self.model.get_filename() is None or isinstance(ex, FileNotFoundError):
+                    ret = False
 
             self._title.set(os.path.basename(filename))
             section_names = self.model.get_section_names()
@@ -235,7 +239,7 @@ class BaseController(ABC):
         pass
 
     def cb_property_select(self, section_name, property_name):
-        """ prpperty selevtion callback """
+        """ property selection callback """
         from qiskit.aqua.parser import JSONSchema
         _show_remove = property_name not in (JSONSchema.PROVIDER, JSONSchema.NAME) \
             if section_name == JSONSchema.BACKEND else property_name != JSONSchema.NAME
@@ -366,7 +370,7 @@ class BaseController(ABC):
         return True
 
     def get_combobox_parameters(self, section_name, property_name):
-        """ get combobox paramaters """
+        """ get combobox parameters """
         from qiskit.aqua.parser import JSONSchema
         values = None
         types = ['string']